chore: move encoding functions into utils package
