+++ huaweicloud/resource_huaweicloud_ecs_instance_v1.go
@@ -1,8 +1,6 @@
 package huaweicloud
 
 import (
-	"crypto/sha1"
-	"encoding/hex"
 	"time"
 
 	"github.com/chnsz/golangsdk"
@@ -56,15 +54,7 @@ func resourceEcsInstanceV1() *schema.Resource {
 				Optional: true,
 				ForceNew: true,
 				// just stash the hash for state & diff comparisons
-				StateFunc: func(v interface{}) string {
-					switch v.(type) {
-					case string:
-						hash := sha1.Sum([]byte(v.(string)))
-						return hex.EncodeToString(hash[:])
-					default:
-						return ""
-					}
-				},
+				StateFunc: utils.HashAndHexEncode,
 			},
 			"password": {
 				Type:      schema.TypeString,
